Remove open contact modal hook
